+++ tests/test_tutorials.py
@@ -89,3 +89,11 @@ def test_iiasa_dbs():
     fname = os.path.join(tut_path, 'iiasa_dbs.ipynb')
     nb, errors = _notebook_run(fname)
     assert errors == []
+
+
+@pytest.mark.skipif(not jupyter_installed, reason=jupyter_reason)
+@pytest.mark.skipif(not pandoc_installed, reason=pandoc_reason)
+def test_ipcc_colors():
+    fname = os.path.join(tut_path, 'ipcc_colors.ipynb')
+    nb, errors = _notebook_run(fname)
+    assert errors == []